Refactor neutrino abstraction
Tux-scripts now has an argument called `target`. This can take the
values "browser", "server" and "multi" and affects what kind of webpack
config is passed to the command.

Passing "browser" or "server" will run the command with a "browser" and
"server" webpack config respectively. Passing "multi" will pass a MultiConfiguration to the command. Note that this may not be supported
by some commands.

It defaults to "browser" except for "start" and "build" which default
to "multi".

This allows us to simplify the neutrino code a bit and create our own
"start" and "build" commands that are registered like other commands.

It also gives an arguably better default for existing or unknown
commands, by default running them only with the "browser" config.
E.g. inspect, test and eslintrc.